Updated READEME, separated the solution and added an implementation interface
* Updated the README with some more useful information and some paragraph improvements.
* Added a solution-folder with some implementation notes.
* Added a shell/implementation interface for the tasks.
