--- src/react-dom.js
@@ -1,1 +1,0 @@
-export { default } from './advanced-react-dom';
