+++ src/react-dom/VCompositeNode.js
@@ -1,0 +1,11 @@
+import { instantiateVNode } from './index';
+
+export default class VCompositeNode {
+    constructor() {}
+
+    getPublicInstance() {}
+
+    update() {}
+
+    mount() {}
+}
