+++ src/react-dom/VDomNode.js
@@ -1,0 +1,11 @@
+import { instantiateVNode } from './index';
+
+export default class VDomNode {
+    constructor() {}
+
+    getPublicInstance() {}
+
+    update() {}
+
+    mount() {}
+}
