+++ src/react-dom/index.js
@@ -1,0 +1,13 @@
+import VCompositeNode from './VCompositeNode';
+import VDomNode from './VDomNode';
+
+export function instantiateVNode() {
+}
+
+function render() {
+}
+
+export default {
+    _reRender: () => render(),
+    render
+};
