+++ src/react/Component.js
@@ -1,0 +1,11 @@
+import ReactDOM from '../react-dom';
+
+class Component {
+    constructor() {}
+
+    setState() {}
+
+    render() {}
+}
+
+export default Component;
