+++ src/react/index.js
@@ -1,0 +1,8 @@
+import Component from './Component';
+
+const createElement = () => ({});
+
+export default {
+    createElement: createElement,
+    Component: Component
+};
