+++ src/solution/README.md
@@ -1,0 +1,4 @@
+# Solution
+
+The proposed solution of ReactDOM and React.
+
