+++ src/solution/react-dom/README.md
@@ -1,0 +1,3 @@
+# ReactDOM
+
+You will find to proposed solutions of ReactDOM with some notes as to how they work in this package.
