+++ src/solution/react-dom/class-cache-react-dom/VCompositeNode.js
