+++ src/solution/react-dom/class-cache-react-dom/VDomNode.js
