+++ src/solution/react-dom/class-cache-react-dom/index.js
