+++ src/solution/react-dom/index.js
@@ -1,0 +1,3 @@
+import ReactDOM from './react-dom';
+
+export default ReactDOM;
