+++ src/solution/react/README.md
@@ -1,0 +1,4 @@
+# React
+
+Here you will find the proposed solution to the implementation of React.
+It is fairly simple with a react-component base class and an index with `React.createElement` in it.
